Invoke theme script inline
